refactor instructions module into class
