+++ lib/jekyll_picture_tag/instructions/preset.rb
@@ -3,8 +3,8 @@ module PictureTag
     # Handles the specific tag image set to construct.
     class Preset
       attr_reader :name
-      def initialize
-        @name = PictureTag.preset_name
+      def initialize(name)
+        @name = name
         @content = build_preset
       end
